Being less stringly typed
Strings don't magically pun as filenames everywhere in Perl 6. Instead
we should use `.IO` to turn a string path into an IO handle. Then we do
the checks on it as methods. The `given` syntax helps here, and we don't
have the `-e` special-case syntax, nor the `_` trick, to remember now.
Note that `.e` is really `$_.e`.