Behaving like a hash
Perl 6 objects are not just hashes underneath. They are an opaque data
structure, efficiently laid out in memory and - post-optimization -
accessed by cheap pointer offsets. But here, we want to behave like a
hash. So we declare a hash attribute, and forward calls on us to the
various hash-y methods onwards to it.